requirements.txt
aiohttp
aiohttp_cors
pyyaml
redis